Revert "target/arm: Revert back to YIELD for WFI"
This reverts commit 5f38ea92fb697b94ad43f01fe162f3ed6e6b0e16.

The commit 5f38ea92fb697b94ad43f01fe162f3ed6e6b0e16 was probably
introduced in order to improve emulation performance, as well as to
address the stability issue when running co-simulations (refer to the
commit 708639f72be2a9425992ae9dc4b9a098dc804c94).

This, unfortunately, breaks the WFI instruction; i.e. CPU is not halted
and keeps running regardless of whether an interrupt occurs, and this
is a problem for the Zephyr CI tests that expect the WFI instruction to
do its job (in particular, tests/kernel/context is broken because of
this change).

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>