Revert "target/arm: Enforce alignment for LDM/STM"
This reverts commit 9d7eb4183a17f129478adbdc314468d84573d83c.

An alignment fault is generated if a data access is performed to an address
that is not aligned for the size of the access, and strict alignment is
required for the access.
https://developer.arm.com/documentation/ddi0460/d/Memory-Protection-Unit/MPU-faults/Alignment-fault?lang=en

The Qemu patch is likely correct and Zephyr test should be fixed but in the
meantime revert this patch to pass tests in CI to give a time to find out
what's wrong with Zephyr test.

When this patch is revert issue is visible by this test:
west twister -v -p qemu_cortex_r5 -O /tmp/twister -T tests/misc/llext-edk/

Signed-off-by: Michal Simek <michal.simek@amd.com>