hw/misc/xlnx-versal-pm[xc]-crp: add reset traces
Add reset related traces to the CRP models of the PMC and PMX
sub-systems. Those traces are:
  - xlnx_versal_pmc_crp_power_on_reset
  - xlnx_versal_pmc_crp_soft_reset

They are triggered on RST_PS writes and report the subsystem (PS, PL or
PMC) for which the reset state is toggled.

Those traces can help developers to catch reset events on the
sub-systems.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>