hw/crypto/xlnx-asu-aes: send a non-zero data mask when CM are on
With DCA counter-measures enabled, when the device send block, it
alternates between a pseudo-random value block and the data block masked
by this value.

Until now the model was sending a block full of zeros as a mask. This
can be an issue if the firmware checks the mask as the counter-measure
become useless with a null mask.

Change this to a constant non-null mask block, which should be enough
for most use-cases.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>